// budget-buddy-ui/src/app/services/models/page-response-expenses-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ExpensesResponse } from '../models/expenses-response';
export interface PageResponseExpensesResponse {
  content?: Array<ExpensesResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}